+++ R/col_by.R
@@ -124,7 +124,7 @@ col_by_to_matrix <- function(col_by, x = NULL) {
     stopifnot(nrow(new_col_by) == `if`(is.data.frame(x), nrow(x), length(x)))
   }
   stopifnot(
-    all(vapply(new_col_by, function(col) is.logical.vector_modif(col, min_size = 0), logical(1)))
+    all_true(new_col_by, is_logical_vector_modif, min_length = 0)
   )
   new_col_by
 }
@@ -243,7 +243,7 @@ by_combine <- function(col_by, ...) {
 #'           FUN = mean, format = "xx.xx")
 #' 
 by_quartile <- function(x, cumulative = FALSE) {
-  stopifnot(is.numeric(x), is.logical.single(cumulative), !any(is.na(x)))
+  stopifnot(is.numeric(x), is_logical_single(cumulative), !any(is.na(x)))
   
   fct <- cut(x, breaks = quantile(x, probs = c(0, 0.25, 0.5, 0.75, 1)), include.lowest=TRUE,
              labels = c("[min, Q1]", "(Q1, Med]", "(Med, Q3]", "(Q3,  max]"))